+++ Competitive.IO/Utf8ConsoleWriter.Span/Utf8ConsoleWriter.cs
@@ -153,6 +153,9 @@ namespace Kzrnm.Competitive.IO
         /// </summary>
         /// <returns>this instance.</returns>
         [M(256 | 512)]
+#if NET9_0_OR_GREATER
+        [O(1)]
+#endif
         public W Write(ReadOnlySpan<char> v)
         {
             var mlen = Utf8NoBom.GetMaxByteCount(v.Length);
@@ -210,6 +213,9 @@ namespace Kzrnm.Competitive.IO
         /// </summary>
         /// <returns>this instance.</returns>
         [M(256)]
+#if NET9_0_OR_GREATER
+        [O(1)]
+#endif
         public W WriteLine(ReadOnlySpan<char> v) { Write(v); return Write('\n'); }
 
         /// <summary>
@@ -238,6 +244,9 @@ namespace Kzrnm.Competitive.IO
         /// Write line each item of<paramref name="col"/>
         /// </summary>
         /// <returns>this instance.</returns>
+#if NET9_0_OR_GREATER
+        [O(1)]
+#endif
         [M(256)] public W WriteLines<T>(ReadOnlySpan<T> col) => WriteMany('\n', col);
 
         /// <summary>
@@ -273,6 +282,9 @@ namespace Kzrnm.Competitive.IO
         /// <returns></returns>
         [M(256)]
         [EditorBrowsable(EditorBrowsableState.Never)]
+#if NET9_0_OR_GREATER
+        [O(1)]
+#endif
         public W WriteMany<T>(char sep, ReadOnlySpan<T> col)
         {
             if (col.Length > 0)