+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2025-01-08
 ### Changed
 - Fix logic
+- Fix document
 
 ## [2.2.0] - 2025-01-02
 ### Added