+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.5.1] - 2025-01-26
+### Changed
+- Fix `ConsoleReader.String`
+### Removed
+- Remove .NET 9 build
+
+
 ## [2.5.0] - 2025-01-26
 ### Added
 - Add `ConsoleReader.Ascii(int defaultBuf)`